+++ .meteor/packages
@@ -15,7 +15,6 @@ es5-shim@4.8.0                # ECMAScript 5 compatibility for older browsers.
 ecmascript@0.14.3              # Enable ECMAScript2015+ syntax in app code
 
 accounts-password@1.6.0
-iron:router@1.1.2
 fourseven:scss@4.12.0
 ejson@1.1.1
 mizzao:timesync@0.5.1
@@ -34,3 +33,5 @@ shell-server@0.5.0
 dynamic-import
 jquery
 mizzao:user-status
+ostrio:flow-router-extra
+kadira:blaze-layout
