+++ .meteor/versions
@@ -41,15 +41,8 @@ htmljs@1.0.11
 http@1.4.2
 id-map@1.1.0
 inter-process-messaging@0.1.1
-iron:controller@1.0.12
-iron:core@1.0.11
-iron:dynamic-template@1.0.12
-iron:layout@1.0.12
-iron:location@1.0.11
-iron:middleware-stack@1.1.0
-iron:router@1.1.2
-iron:url@1.1.0
 jquery@1.11.11
+kadira:blaze-layout@2.3.0
 lai:collection-extensions@0.2.1_1
 launch-screen@1.2.0
 livedata@1.0.18
@@ -75,6 +68,7 @@ npm-bcrypt@0.9.3
 npm-mongo@3.7.0
 observe-sequence@1.0.16
 ordered-dict@1.1.0
+ostrio:flow-router-extra@3.7.4
 peerlibrary:assert@0.3.0
 peerlibrary:fiber-utils@0.10.0
 peerlibrary:server-autorun@0.8.0
