+++ app/admin/client/accounts.html
@@ -1,4 +1,5 @@
 <template name="accounts">
+{{#with accounts}}
   {{#unless empty unapproved_players}}
     <h2>Unapproved Players</h2>
     <table class="table table-striped">
@@ -64,4 +65,5 @@
       </tbody>
     </table>
   {{/unless}}
+{{/with}}
 </template>
