+++ app/authentication/client/approval.html
@@ -1,3 +1,3 @@
 <template name="approval">
-  Your account needs to be approved before you can play
+  <div>Your account needs to be approved before you can play</div>
 </template>
