+++ app/authentication/client/approval.js
@@ -1,0 +1,11 @@
+import { FlowRouter } from 'meteor/ostrio:flow-router-extra'
+
+Template.approval.onCreated(registerStreams)
+
+function registerStreams() {
+  Streamy.on('account_approved', redirectToLobby)
+}
+
+function redirectToLobby(data) {
+  FlowRouter.go(`/lobby`)
+}
