+++ app/game/client/templates/game.html
@@ -1,5 +1,5 @@
 <template name="game">
-
+{{#with current_game}}
   <div class="row" id="game">
 
     <div class="col" id="cards-area">
@@ -110,5 +110,5 @@
       </div>
     </div>
   </div>
-
+{{/with}}
 </template>
