+++ app/layout/client/head.html
@@ -1,0 +1,4 @@
+<head>
+  <link rel="icon" type="image/png" href="/favicon.png">
+  <title>Dominion</title>
+</head>
