+++ app/layout/client/layout.html
@@ -1,7 +1,3 @@
-<head>
-  <link rel="icon" type="image/png" href="/favicon.png">
-  <title>Dominion</title>
-</head>
 <template name="layout">
   <nav class="navbar navbar-expand-xl navbar-dark bg-dark fixed-top">
     <a class="navbar-brand" href="/">Dominion</a>
