+++ app/layout/client/layout.js
@@ -1,9 +1,13 @@
-Template.layout.events({
+import { FlowRouter } from 'meteor/ostrio:flow-router-extra'
 
+Template.layout.events({
   'click .js-logout': function() {
     Meteor.logout(function() {
-      Router.go('/')
+      FlowRouter.go('/')
     })
+  },
+  'click .nav-link, click .navbar-brand': function(event) {
+    event.preventDefault()
+    FlowRouter.go(event.target.getAttribute('href'))
   }
-
 })
