+++ app/players/client/templates/players_show.html
@@ -1,4 +1,5 @@
 <template name="playersShow">
+{{#with player}}
   <h2>{{username}}</h2>
   <h3>Win Ratio: {{win_ratio}}%</h3>
   <h3>Wins: {{wins}}</h3>
@@ -62,4 +63,5 @@
     </tbody>
   </table>
   {{> pagination}}
+{{/with}}
 </template>
