+++ lib/client/template_helpers.js
@@ -1,3 +1,5 @@
+import { FlowRouter } from 'meteor/ostrio:flow-router-extra'
+
 Template.registerHelper('times', function (times) {
   if (times !== undefined) {
     return _.times(times, function(counter) {
@@ -21,7 +23,7 @@ Template.registerHelper('unstyled_username', function (player) {
 })
 
 Template.registerHelper('my_turn', function () {
-  let game = Games.findOne(Router.current().params.id)
+  let game = Games.findOne(FlowRouter.getParam('id'))
   if (game.turn.possessed) {
     return game.turn.possessed._id === Meteor.userId()
   } else {
@@ -30,17 +32,17 @@ Template.registerHelper('my_turn', function () {
 })
 
 Template.registerHelper('allow_villagers', function () {
-  let game = Games.findOne(Router.current().params.id)
+  let game = Games.findOne(FlowRouter.getParam('id'))
   return _.includes(['action'], game.turn.phase)
 })
 
 Template.registerHelper('allow_treasures', function () {
-  let game = Games.findOne(Router.current().params.id)
+  let game = Games.findOne(FlowRouter.getParam('id'))
   return _.includes(['action', 'treasure'], game.turn.phase)
 })
 
 Template.registerHelper('allow_debt', function () {
-  let game = Games.findOne(Router.current().params.id)
+  let game = Games.findOne(FlowRouter.getParam('id'))
   return _.includes(['action', 'treasure', 'buy'], game.turn.phase)
 })
