--- lib/controllers/core/application_controller.js
@@ -1,4 +1,0 @@
-ApplicationController = RouteController.extend({
-  layoutTemplate: 'layout',
-  loadingTemplate: 'loading'
-})
