--- lib/controllers/core/logged_in_admin_controller.js
@@ -1,13 +1,0 @@
-LoggedInAdminController = ApplicationController.extend({
-
-  onBeforeAction: function () {
-    if (!Meteor.userId()) {
-      this.redirect('login')
-    } else if (!Meteor.user().admin) {
-      this.redirect('lobby')
-    } else {
-      this.next()
-    }
-  }
-
-})
