--- lib/controllers/forgot_password_controller.js
@@ -1,2 +1,0 @@
-ForgotPasswordController = AuthenticationController.extend({
-})
