--- lib/controllers/login_controller.js
@@ -1,2 +1,0 @@
-LoginController = AuthenticationController.extend({
-})
