--- lib/controllers/reset_password_controller.js
@@ -1,7 +1,0 @@
-ResetPasswordController = AuthenticationController.extend({
-
-  onBeforeAction: function() {
-    Accounts._resetPasswordToken = this.params.token
-    this.next()
-  }
-})
