--- lib/controllers/signup_controller.js
@@ -1,2 +1,0 @@
-SignupController = AuthenticationController.extend({
-})
